assetmng-main/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
pydantic
python-multipart
python-dotenv